### 2020-01-20 #2031 @whedon
[PRE REVIEW]: MF2: A Collection of Multi-Fidelity Benchmark Functions in Python
**Submitting author:** @sjvrijn (<a href="http://orcid.org/0000-0001-6159-041X">Sander van Rijn</a>)
**Repository:** <a href="https://github.com/sjvrijn/mf2" target ="_blank">https://github.com/sjvrijn/mf2</a>
**Version:** 2019.11.3
