### 2020-01-20 e227ecd
Added JOSS badge to README
