### 2020-01-21 d8fc58b
adjustable functions: bounds now separate variables, docstrings updated with correct parameter values for high-low equality
