+++ mf2/adjustable/paciorek.py
@@ -34,18 +34,22 @@ def adjustable_paciorek_lf(xx, a2):
     return temp1 - (temp2*temp3)
 
 
+u_bound = [1]*2
+l_bound = [0.3]*2
+
+
 def paciorek(a2):
     """Factory method for adjustable Paciorek function using parameter value `a2`
 
     :param a2:  Parameter to tune the correlation between high- and low-fidelity
                 functions. Expected values lie in range [0, 1]. High- and low-
-                fidelity are identical for a1=-0.5.
+                fidelity are identical for `a2=0.0`.
     :return:    A MultiFidelityFunction instance
     """
 
     return MultiFidelityFunction(
         f"adjustable Paciorek {a2}",
-        [1]*2, [0.3]*2,
+        u_bound, l_bound,
         [paciorek_hf, partial(adjustable_paciorek_lf, a2=a2)],
         fidelity_names=['high', 'low'],
     )
