### 2020-01-29 #2031 @petroniocandido
@melissawm Yes, sure! It is my first review for FOSS,  is chefe a manual of good practices?
