### 2020-01-30 #2031 @melissawm
@petroniocandido you can check the general reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and if you have any questions, please let me know.

Once I've assigned you as reviewer, this issue is going to be closed and a review issue will be opened with a checklist for each of the reviewers.
