### 2020-01-30 #2049 @whedon
[REVIEW]: MF2: A Collection of Multi-Fidelity Benchmark Functions in Python
**Submitting author:** @sjvrijn (<a href="http://orcid.org/0000-0001-6159-041X">Sander van Rijn</a>)
**Repository:** <a href="https://github.com/sjvrijn/mf2" target ="_blank">https://github.com/sjvrijn/mf2</a>
**Version:** v2020.8.0

@petroniocandido, @torressa please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
