### 2020-02-17 #2049 @melissawm
Hello @jgoldfar, would you be willing/available to review this software paper for JOSS? Thank you
