### 2020-03-06 #2049 @torressa
Hi @sjvrijn! Looks like you've put some work into the package!

I'm not that familiar with evolutionary optimisation algorithms, so please mind the ignorant comments.

I have the following comments:

1. The pip installation doesn't automatically install the required dependencies. (tested on a clean pyenv).
2. Also, when trying to run `python3 tests/create_regression_data.py` I get `ImportError: attempted relative import with no known parent package`.
3. You mention in the paper, the `numbbo/coco` package. Their package is pretty well established, documented, and tested.
I've seen that the functions you've implemented are not present, so my question is,
why did you decide to release your own package instead of contributing?
I think you need to explain, preferably also in the paper,
how your package compares to the other(s) and why you choose to release your work separately.
4. Code-wise:
    - I see there's a fair amount of duplicated code, particularly, the `artificialMultifidelity.py` and `borehole.py` (from codacy).
    - Some contributing guidelines would be good.
    - I've got a comment about performance, in your functions, everything is done using vector/matrix manipulations,
even though they're pretty efficient in other programming languages, for python, as far as I understand it,
it's not usually the fastest way. A quick google search brings up some [examples](https://zerowithdot.com/python-numpy-and-pandas-performance/) where list comprehensions are better than their numpy equivalent.
However, this requires a major overhaul of the package for a (potentially) small increase in performance,
so not expecting that.
Would be cool though to have some results with increasingly larger instances to see how the implementations scale.
5. I like the docs, there's explanations, simple examples and how to customise the `MultiFidelityFunction` class.
You mention in the paper that you're using
the package for your research, so it'd be good if you could provide a more in-depth example (preferably also with the code).
I think the API docs need a bit of work, the docs `MultiFidelityFunction` are good but the rest of the functions
could be improved.
6. As I said, I'm not familiar with the topic, but are there some benchmark instances to which you can compare
your implementations with? Perhaps against the matlab counterparts?
You could maybe add this as the in-depth example and kill several birds with one stone.

That's all I've got for now, keep up the good work!
