### 2020-03-07 #2049 @sjvrijn
Hi @torressa, thanks for the compliments and comments!

1. I realised this recently, will add it asap.
2. I've done most of my development from PyCharm, which tends to run files as-is. Not the first time I've encountered import issues because of it, but hadn't seen this one yet, so thanks for pointing that out 👍
3. That's a fair question to ask. I'd say it boils down to a difference in the kinds of functions that are used which wouldn't nicely intergrate with `coco`. I'll update the paper with a more complete comparison and explanation.
4. - `artificialMultifidelity` is semi-dead code at the moment, so you're right, I should properly include or remove it. For `borehole.py` there are some subtle differences which it doesn't highlight, so not de-duplicating has been a concious decision so far. I'll have a look at it again to see if that's still the right decision.
   - Can you be more specific? Anything specific you would like to see added to the [Contributing section](https://github.com/sjvrijn/mf2#contributing) of the readme?
   - Matrix/vector operations in numpy are passed through to C/fortran procedures, so as long as the amount of calls is kept to a minimum, they are much faster than pure python loops/comprehensions from what I have experienced. I believe the examples you linked to show the same thing.
     Meanwhile, the typical evaluation budget is very low (~1000 evals.) in the situations where multi-fidelity optimization becomes relevant.
     Either way, thanks for suggesting the inclusion of a plot to show how the implementation scales, I agree that can be a useful addition.
5. I'll see what I can do for a more in-depth but still stand-alone example
6. Apart from a few implementations [here](https://www.sfu.ca/~ssurjano/multi.html), I don't know of any other public implementations (which is one of the main reasons I wrote this package in the first place 😅 ).

I'll give a notice when I've updated anything.

@melissawm Could you edit the fist post by Whedon to add a check-list for @torressa? Would help me to see what I still have to work on too 🙂
