### 2020-03-12 #2049 @torressa
Hi @sjvrijn! Sorry for the delay and thanks for the reply.
I've updated the check point list :sunglasses:
Thanks @melissawm

4.2 That looks good actually, missed the subsections on my first look. Also, pretty cool that you set up a gitter channel.
4.3 Fair enough, didn't know that. I just came across several cases when python constructs (maps, generators and list comprehensions) where a lot faster than numpy. I think with the appropriate construct, they are comparable to numpy both when the number of calls is large and when the size of the arrays is also large (which is what happens in all the 2D examples in the link in my original comment, and that's with list comprehensions!).
The scalability plot would be great!
6. Oh cool, I think if the implementations are fairly well known in the field, a simple comparison could help convince people to use your package. It should be fairly easy if you have access to a matlab license.

As for referencing the updates, I'll raise some issues in your repo with the relevant changes. If you  add `#issue_number` to your commit message, github will automatically show it here.
