### 2020-03-12 sjvrijn/mf2#1 @torressa
JOSS Review. Repo Changes.
According to my [review](https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-595979532) and @sjvrijn  [response](https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-596131480).

The following need work:

 - [x] The pip installation doesn't automatically install the required dependencies. (tested on a clean pyenv).
 - [x] Also, when trying to run `python3 tests/create_regression_data.py` I get `ImportError: attempted relative import with no known parent package`.
 - [x] I see there's a fair amount of duplicated code, particularly, the `artificialMultifidelity.py` and `borehole.py` (from codacy).
 - [x] Scalabiliy plot for implementations.
 - [x] (lf possible) I think if the [implementations](https://www.sfu.ca/~ssurjano/multi.html) are fairly well known in the field, a simple comparison could help convince people to use your package. It should be fairly easy if you have access to a matlab license.

_Originally posted by @torressa in https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-595979532_
and
_Originally posted by @torressa in https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-598033119_
