### 2020-03-12 sjvrijn/mf2#2 @torressa
JOSS Review. Paper Changes.
According to my [review](https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-595979532) and @sjvrijn  [response](https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-596131480).

The following needs work:
 - [x] Explanation in the paper how your package compares to the other(s) and (if relevant) why you choose to release your work separately.

_Originally posted by @torressa in https://github.com/openjournals/joss-reviews/issues/2049#issuecomment-595979532_
