### 2020-03-18 c31559e
regular requirements recursively installed by requirements-dev; README includes requirements-dev install instruction
