### 2020-03-18 0718969
Version number stored in separate file; cleaned bi_fidelity_functions definition in __init__
