### 2020-03-18 bce30cb
Updated setup: version now read from separate file, not from module; added classifiers
