### 2020-03-19 ba94b4c
version sourcing changed back to setup.py; __init__ now uses pkg_resources to retrieve it
