### 2020-03-19 e09053b
header cleanup of tests.utils.py
