### 2020-03-19 6ee6b3c
cleared tests/__init__.py
