+++ tests/__init__.py
@@ -1,13 +1,0 @@
-#!/usr/bin/python3
-# -*- coding: utf-8 -*-
-
-"""
-Filename.py: << A short summary docstring about this file >>
-"""
-
-__author__ = 'Sander van Rijn'
-__email__ = 's.j.van.rijn@liacs.leidenuniv.nl'
-
-
-
-
