+++ setup.py
@@ -1,5 +1,4 @@
 import setuptools
-from pathlib import Path
 
 with open('README.md') as f:
     long_description = f.read()
