### 2020-03-19 dfe522f
refactored path definitions in tests/create_regression_data.py
