### 2020-03-19 bdb5f84
updated README on how to run 'create_regression_files' script
