### 2020-03-19 ab87a2b
making travis actually install the package
