### 2020-03-22 #2049 @sjvrijn
First of all, I completely understand it if any of you have more pressing matters than this review given the current circumstances. So please take care of yourself and many thanks for any time you are still able to contribute to this review.

@torressa While creating the scalability plot (coming soon), as a reply to your concerns about performance of using Numpy, I've compared a pre-numpy implementation of the Hartmann6 function with the numpy implementation. As you can see from the table below, the numpy implementation is ~10-80 times faster for evaluating N>=10 points at a time.

|       N | Pure Python |    Numpy | Speedup |
|--------:|------------:|---------:|--------:|
|       1 |    0.000026 | 0.000015 |     1.7 |
|      10 |    0.000229 | 0.000022 |    10.5 |
|     100 |    0.002211 | 0.000045 |    49.3 |
|    1000 |    0.020856 | 0.000279 |    74.7 |
|   10000 |    0.197703 | 0.005815 |    34.0 |
|  100000 |    2.094764 | 0.047761 |    43.9 |
| 1000000 |   20.371699 | 0.515337 |    39.5 |

For completeness: I've taken the code from commit `f22104d3e62a4535f660cad73b8a822cc1f3cef0` and adjusted it. You could test this yourself by adding the code below into the `mf2/hartmann.py` file

```python
import math

A = _A6[0]
P = _P6[0]
alpha_high = _alpha6_high.T[0]
alpha_low = _alpha6_low.T[0]

def pure_python_hartmann6_hf(xx):
    xx = np.atleast_2d(xx)
    results = []
    for row in xx:
        outer_sum = 2.58
        for alpha_i, A_i, P_i in zip(alpha_high, A, P):
            inner_sum = 0
            for xx_j, A_ij, P_ij in zip(row, A_i, P_i):
                inner_sum += A_ij * (xx_j-P_ij)**2

            outer_sum += alpha_i * math.exp(-inner_sum)
        results.append(-(1/1.94) * outer_sum)
    return np.array(results)
```
