### 2020-03-22 979ef3b
Added scalability comparison figure to JOSS paper (addressess point 4 of #1)
