### 2020-03-24 #2049 @torressa
Hi @sjvrijn!
Cool scalability plot, and pretty neat code to create it as well.
About the numpy vs pure python discussion, thanks for making the extra effort.
I implemented a faster pure python version with generators and maps but was still slower than numpy. You win!
