### 2020-03-24 sjvrijn/mf2#1 @torressa
The installation doesn't atomatically install numpy.
I think maybe you need to release the changes to pypi first.
I did
```none
python3 -m venv env
. env/bin/activate
pip install mf2
```
and I get
```none
Collecting mf2
  Downloading https://files.pythonhosted.org/packages/d0/43/5a6e4f2c33e36776eb3a7b7cab13cd045d84d9514c298d220bccbcc79088/mf2-2019.11.3-py3-none-any.whl
Installing collected packages: mf2
Successfully installed mf2-2019.11.3
```
