### 2020-03-24 sjvrijn/mf2#1 @sjvrijn
> I think maybe you need to release the changes to pypi first.
True, I've only updated the source, not yet the PyPI version
