+++ mf2/forrester.py
@@ -29,14 +29,15 @@ def forrester_high(X):
     return np.sum(term1 * term2, axis=1) / ndim
 
 
-def forrester_low(X):
+def forrester_low(X, *, A=0.5, B=10, C=-5):
     X = np.atleast_2d(X)
 
     ndim = X.shape[1]
-    term1 = 0.5*forrester_high(X)
-    term2 = 10*(X - 0.5) - 5
+    term1 = A*forrester_high(X)
+    term2 = B*(X - 0.5)
+    term3 = C
 
-    return term1 - (np.sum(term2, axis=1) / ndim)
+    return term1 + (np.sum(term2, axis=1) / ndim) + term3
 
 
 #: Lower bound for Forrester function
