### 2020-04-05 1755990
Added docs/example-usage: recreating fig. 1 from forrester 2007
