### 2020-04-07 df27362
Added MANIFEST.in file to include requirements and docs in source distribution
