### 2020-04-07 89086cb
removed artificialmultifidelity.py
