+++ mf2/artificialMultifidelity.py
@@ -1,147 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-artificialMultifidelity.py:
-
-This function is a based on a set of equations with one-dimensional
-input and output. The lowest fidelity is 1, while 6 is the highest
-fidelity available. Exact formulas as given in the original paper by
-J. Branke et al. are listed below. For input in more than one dimension,
-the sum over all values as calculated for each dimension is returned. Note
-that the location of the global optimum shifts from the right for fidelities
-1-2, to equal for fidelities 3-4 and finally to the left for fidelities 5-6.
-This is on purpose to simulate the occurrence of local optima in lower
-fidelity evaluations.
-
-Below are the exact formulas for all 6 fidelity levels.
-
-\f{eqnarray*{ \\
-f1(x) =& \min \{ &
-         (x-2)^2,\\
-      && (x+2)^2 + 2 \} \\
-
-f2(x) =& \min \{ &
-         (x-2)^2 + 5\sin ( \frac{\pi}{2} (x+1)),\\
-      && (x+2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + \frac{6}{5} \} \\
-
-f3(x) =& \min \{ &
-         (x-2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )),\\
-      && (x+2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        \frac{2}{5} \} \\
-
-f4(x) =& \min \{ &
-        (x-2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )),\\
-     && (x+2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )) - \frac{2}{5} \} \\
-
-f5(x) =& \min \{ &
-        (x-2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )) + 2\sin ( 4\pi (x + \frac{15}{8} )), \\
-     && (x+2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )) + 2\sin ( 4\pi (x + \frac{15}{8} )) -
-        \frac{6}{5} \} \\
-
-f6(x) =& \min \{ &
-        (x-2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )) + 2\sin ( 4\pi (x + \frac{15}{8} )) + \sin ( 8\pi (x+2)), \\
-     && (x+2)^2 + 5\sin ( \frac{\pi}{2} (x+1)) + 4\sin ( \pi (x + \frac{3}{2} )) +
-        3\sin ( 2\pi (x + \frac{7}{4} )) + 2\sin ( 4\pi (x + \frac{15}{8} )) +
-        \sin ( 8\pi (x+2))- 2 \} \\
- \f}
-
-References:
-
-[Efficient Use of Partially Converged Simulations in Evolutionary Optimization, J. Branke et al. 2016.]
[…]
