+++ changelog.txt
@@ -1,5 +1,5 @@
 since last version:
-
+- Removed unused artificialMultiidelity.py
 
 v2020.4.1:
 - Added Manifest.in to include requirements and docs in source distribution
