### 2020-04-07 311450c
De-duplicated code for borehole fidelities
