### 2020-04-07 4667267
Added explicit statement of need paragraph
