### 2020-04-07 701da09
scalability-comparison now using fixed seed; updated figure
