### 2020-04-07 91b5fdb
Fixed typo in docs: 'Burrin' -> 'Currin'
