### 2020-04-08 4884708
Added LICENSE to MANIFEST.in
