+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include requirements.txt requirements-dev.txt
+include LICENSE requirements.txt requirements-dev.txt
 recursive-include docs *.rst
