### 2020-04-08 0b60263
Docs: added 'Adjustable' to titles of adjustable functions
