+++ docs/source/functions/adjustable/trid.rst
@@ -1,5 +1,5 @@
-Trid
-====
+Adjustable Trid
+===============
 
 .. automodule:: mf2.adjustable.trid
     :members:
