+++ docs/source/functions/adjustable/hartmann.rst
@@ -1,5 +1,5 @@
-Hartmann3
-=========
+Adjustable Hartmann3
+====================
 
 .. automodule:: mf2.adjustable.hartmann
     :members:
