+++ docs/source/functions/adjustable/paciorek.rst
@@ -1,5 +1,5 @@
-Paciorek
-========
+Adjustable Paciorek
+===================
 
 .. automodule:: mf2.adjustable.paciorek
     :members:
