### 2020-04-08 a93b105
docs: updated copyright and develop install instructions
