### 2020-04-09 4b3c8d0
Bugfix: corrected upper/lower bounds for adjustable Trid function
