### 2020-04-12 244d400
scalability plot separated into function
