+++ JOSS-paper/scalability-comparison.py
@@ -39,6 +39,41 @@ def create_and_store_time_scaling_data():
     return df
 
 
+def plot_mf2_scalability(df):
+
+    fig, axes = plt.subplots(1, 2, figsize=(9.6, 4.8), constrained_layout=True)
+    plt.suptitle('Scalability of mf2-functions')
+    xticks = pd.unique(df['size'])
+    for ax, (fid, df_per_fid) in zip(axes, df.groupby('fidelity')):
+        ax.set_xscale('log')
+        ax.set_yscale('log')
+        ax.set_xlabel('$N$')
+        ax.set_ylabel('$t/t_0$')
+        ax.set_title(f'{fid} fidelity')
+        for ndim, df_per_dim in df_per_fid.groupby('ndim'):
+            data = []
+            for name, sub_df in df_per_dim.groupby('name'):
+                data.append(sub_df['norm_time_per'].values)
+            data = np.array(data)
+
+            mean = np.mean(data, axis=0)
+            label = f'{ndim}D'
+            if data.shape[0] == 1:
+                ax.plot(xticks, mean, label=label)
+            else:
+                min_max = np.abs(np.array([mean - np.min(data, axis=0),
+                                           np.max(data, axis=0) - mean]))
+                ax.errorbar(xticks, mean, min_max, label=label, capsize=4)
+
+
+    axes[0].legend(loc=0)
+    axes[1].legend(loc=0)
+    plt.savefig('scalability.pdf')
+    plt.savefig('scalability.png')
+    plt.show()
+
+
+np.random.seed(20160501)
 save_location = Path('time_scaling.csv')
 if not save_location.exists():
     df = create_and_store_time_scaling_data()
@@ -46,35 +81,4 @@ if not save_location.exists():
 else:
     df = pd.read_csv(save_location)
 
-
-np.random.seed(20160501)
-fig, axes = plt.subplots(1, 2, figsize=(9.6, 4.8), constrained_layout=True)
-plt.suptitle('Scalability of mf2-functions')
-xticks = pd.unique(df['size'])
-for ax, (fid, df_per_fid) in zip(axes, df.groupby('fidelity')):
-    ax.set_xscale('log')
-    ax.set_yscale('log')
-    ax.set_xlabel('$N$')
-    ax.set_ylabel('$t/t_0$')
-    ax.set_title(f'{fid} fidelity')
-    for ndim, df_per_dim in df_per_fid.groupby('ndim'):
-        data = []
-        for name, sub_df in df_per_dim.groupby('name'):
-            data.append(sub_df['norm_time_per'].values)
-        data = np.array(data)
-
-        mean = np.mean(data, axis=0)
-        label = f'{ndim}D'
-        if data.shape[0] == 1:
-            ax.plot(xticks, mean, label=label)
-        else:
-            min_max = np.abs(np.array([mean - np.min(data, axis=0),
-                                       np.max(data, axis=0) - mean]))
-            ax.errorbar(xticks, mean, min_max, label=label, capsize=4)
-
-
-axes[0].legend(loc=0)
-axes[1].legend(loc=0)
-plt.savefig('scalability.pdf')
-plt.savefig('scalability.png')
-plt.show()
+plot_mf2_scalability(df)
