### 2020-04-12 f8584cd
Added function for scalability comparison plot: mf2 vs surjanovic/bingham matlab implementation
