### 2020-04-12 4b6aea4
moved scalability plot to docs/_static
