### 2020-04-12 154b8a7
docs/example-usage updated: code extracted to scripts/example-usage.py; figure upgraded to 600dpi
