### 2020-04-12 1a3f5fd
added performance page to docs, includes scalability plots
