### 2020-04-12 fb909e6
location of docs/_static updated in sphinx conf.py
