+++ changelog.txt
@@ -1,4 +1,8 @@
 since last version:
+- Documentation:
+  * example-usage separated out into separate file in docs/scripts
+  * figure of example-usage improved to 600dpi
+  * page on performance added, script for plots added in docs/scripts
 
 
 v2020.4.3:
