### 2020-04-12 #2049 @sjvrijn
Hi @torressa, I believe I've now addressed all the issues you've raised in https://github.com/sjvrijn/mf2/issues/1 and https://github.com/sjvrijn/mf2/issues/2:

- I've added a documentation page: [Performance](https://mf2.readthedocs.io/en/latest/performance.html) which includes a comparison to the matlab code by [Surjanovic and Bingham](https://www.sfu.ca/~ssurjano/multi.html),
- added a statement of need paragraph to the paper (5th paragraph "This new class of ..."),
- reduced duplicated code in `borehole.py` (and removed `artificialmultifidelity.py` for the time being),
- updated contribution instructions with a correct command for running the `create_regression_files.py` script (`python3 -m test.create_regression_files`),
- and created a new PyPI release that does automatically install the requirements.
