### 2020-04-17 dea861c
updated badges: travis-ci.org -> .com; conda|sjvrijn -> conda-forge
