### 2020-04-17 c603232
coverage reporting switched from using 'python-coveralls' to just 'coveralls' in Travis-CI
