+++ .travis.yml
@@ -10,7 +10,7 @@ matrix:
 install:
   - pip install .
   - pip install -r requirements-dev.txt
-  - pip install coverage python-coveralls
+  - pip install coverage coveralls
 script:
   - pytest tests/regression_test.py
   - pytest --cov=mf2 tests/property_test.py
