### 2020-04-17 d7774f6
updated install instructions in README
