### 2020-05-20 #2049 @arfon
:wave: @jmadera and @zbeekman - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
