### 2020-05-20 #2049 @arfon
> I'd love to! Looks interesting!

Great, thanks so much @zbeekman! I've added a checklist for you at the top of this issue. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
