### 2020-05-20 #2049 @zbeekman
@arfon have the reviewer guidelines changed meaningfully in the past ~3-6 months? I'll skim them again to make sure I don't miss anything but I've reviewed a few JOSS papers before.
