### 2020-05-20 #2049 @zbeekman
Hi @sjvrijn! Both the paper and package are looking pretty great from my initial inspection!

For issues specific to documentation, installation, etc. (package stuff) I'll use issues on the project's github repo, but, unless you object, I'd like to post my comments about the paper here. (To me the project repository feels like the wrong place for that.)

In general I think the paper is great, but there are a few small things that I think should be addressed.

1. __Statement of need__: I think paragraphs 4 and 5 really demonstrate the need or hole your package is filling, but I think this could be made clearer in the following way:
   - Paragraph 4 is quite short and starts with: "Because multi-fidelity problems are a new class of problems, the existing single-fidelity benchmark suites that exist cannot be used for this field." However it is not immediately clear *why* this is the case (to a non-expert, at least). Paragraph 5 gives a fuller answer to this question. I think paragraphs 4 and 5 can easily be combined with minimal or no restructuring or rewording. (At most you could shuffle one or two sentences around.) This is not a mandatory change, but the opening sentence of paragraph 4 felt unsubstantiated and left me questioning it until I had read through the first three sentences of paragraph 5.
2. __Scalability plot/discussion__: Normalizing by N=1 time is not enough. I think the following information needs to be explicitly included:
   - Hardware/CPU info
     - Different architectures have different performance characteristics, without this info it's hard to reproduce the findings.
   - Runtime software info (python version; numpy, MF2, etc. versions)
   - Parallelism information: Is the simulation run on a single thread? Multiple threads? Please be explicit about this.
   - If each series uses a different function, it would be nice to add additional abbreviations to the series labels: the label `<N>D` makes it seem like only the dimensionality of the problem is changing
   - Please ensure the abscissa (looks ok) and ordinate (needs fixing) of the two plots (high vs low fidelity) are the same. In log coordinates, especially when auto-scaling the range of the ordinate, it appears that the low fidelity runs take exactly the same amount of time (or scale exactly the same) as the high fidelity runs.
   - Please clarify if the `N = 1` ( `t_0` ) normalization is the same or different between the high fidelity and low fidelity plots.
3. `multi- fidelity` in the last sentence of prose before the acknowledgements section has a strange looking space. I'm not sure if this is due to JOSS rendering or an issue in the paper's markdown source.
4. An additional final sentence or clause, perhaps explaining how/why the authors are using the package in their research or something summarizing the package capabilities might help make the ending feel less abrupt and tie things together.

IMO, these are pretty much nits, except for 2. (scalability plot/discussion). I think 1, 3 and 4 would improve the paper but welcome push back and explanation if you disagree.

Once you respond and/or make changes should be able to close out the other checkboxes for the paper section of my review.

I know you've been waiting quite a while now, so I'll try to be prompt with my review of documentation, installation, etc. but I will need to perform that later tonight, tomorrow or the day after.
