### 2020-05-21 59ac7d6
JOSS paper: fixed mid-word space
