### 2020-05-21 b51fe62
JOSS paper: rewrote paragraphs 4&5 for clarity
