### 2020-05-22 6bfa71b
JOSS paper: better description of usage by authors
