### 2020-05-22 982f6ea
name change: mf2.adjustable.bi_fidelity_functions now consistenw with mf2.bi_fid...
