### 2020-05-22 7294860
cleaned up scalability-comparison.py: script now in a main()
