### 2020-05-22 8093e0c
Docs: added hardware and version information to performance page
