+++ docs/source/performance.rst
@@ -10,7 +10,8 @@ programs.
 
 The scripts for generating following performance overviews can be found in the
 `docs/scripts <https://github.com/sjvrijn/mf2/tree/master/docs/scripts>`_ folder
-of the repository.
+of the repository. Presented running times were measured on a desktop PC with an
+Intel Core i7 5820k 6-core CPU, with Python 3.6.3 and Numpy 1.18.4.
 
 Performance Scaling
 -------------------
@@ -20,9 +21,11 @@ functions simultaneously as a matrix of size ``(N, ndim)``. Performance for the
 high- and low-fidelity formulations are shown separately to give a fair
 comparison: many low-fidelities are defined as computations on top of the
 high-fidelity definitions. As absolute performance will vary per system, the
-runtime is divided by the time needed for ``N=1`` as a normalization. Up to
-``N=1_000``, the time required scales less than linearly thanks to efficient
-and vectorized numpy routines.
+runtime is divided by the time needed for ``N=1`` as a normalization. This is
+done independently for each function and fidelity level.
+
+Up to ``N=1_000``, the time required scales less than linearly thanks to
+efficient and vectorized numpy routines.
 
 .. image:: ../_static/scalability.png
   :width: 640
@@ -35,7 +38,8 @@ The following image shows how the scaling for the ``mf2`` implementation of the
 **Currin**, **Park91A**, **Park91B** and **Borehole** functions compares to the
 *Matlab* implementations by `Surjanovic and Bingham
 <https://www.sfu.ca/~ssurjano/multi.html>`_, which can only evaluate one point
-at a time, so do not use any vectorization.
+at a time, so do not use any vectorization. Measurements were performed using
+*Matlab* version R2020a (9.8.0.1323502).
 
 .. image:: ../_static/scalability_comparison.png
   :width: 640
