### 2020-05-22 4f59fcd
scaling data now includes adjustable bi-fidelity functions at A=0.5
