### 2020-05-22 ca7b275
Improved clarity of scalability plot labels
