### 2020-05-22 601dc29
Explicitly fixed upper x/y limits
