### 2020-05-22 e6e1c8a
Added matlab performance testing scripts + results csv
