### 2020-05-22 d112238
JOSS paper: Added hardware/software version information
