### 2020-05-22 e4ed1de
JOSS paper: added explicit reference to using numpy for parallelism
