### 2020-05-22 #2049 @sjvrijn
Hi @zbeekman! Thanks for the prompt review, I appreciate it and will do my best to reciprocate with my own responses :)

I've updated the paper (and performance page of the docs) to respond to just about all of your comments. Regarding your question about parallelism, the package simply relies on numpy and doesn't perform any explicit parallelization itself. I've tried to look whether it ran on one or more cores, but I'm not sure to be honest 😅 For some functions it clearly only uses one thread in task manager, but for others it's chaotic. So I haven't added any explicit mention of parallelism, but have listed core count and stated in the paper that any parellelism comes from numpy (which in turn depends on the BLAS/LAPACK backend). Does that satisfy your comment?
