### 2020-05-26 #2049 @zbeekman
Hi @sjvrijn,
Thanks for addressing the substance of my comments about the paper so well! The changes look great, and I (personally) think the statement of need is clearer, especially to non-experts. The wording in that new/revised paragraph P4 could benefit from a few small edits to reduce redundancy and improve flow. I'll open a PR with my suggestions later today, nothing serious, just very small edits. Beyond that, as far as I'm concerned, the paper looks really good.

I'll do my best to get to the technical, usage and documentation review this afternoon along with my small wording suggestions.
