### 2020-05-30 49b723d
Refactored by Sourcery
