### 2020-06-12 b13585d
Replaced example function used in README
