### 2020-06-12 ff399cd
forcing travis to update to most recent dependency versions
