### 2020-06-21 d9a47a3
removed 3-fidelity himmelblau versions
