+++ mf2/himmelblau.py
@@ -29,33 +29,6 @@ def himmelblau_hf(xx):
     return term1 + term2
 
 
-def himmelblau_mf(xx):
-    """
-    HIMMELBLAU FUNCTION, MEDIUM FIDELITY CODE
-
-    INPUT:
-    xx = [x1, x2]
-    """
-    xx = np.atleast_2d(xx)
-
-    x1, x2 = xx.T
-
-    x1 *= .75
-    x2 *= .9
-
-    term1 = (x1**2 + x2 - 11)**2              # A -- E
-    # term2 = 7*x2                              # A
-    # term2 = 0                                 # B
-    # term2 = (x2**2 - 7)**2                    # C
-    # term2 = (x2**2 - 7)**2 + 5*x2**2 - 28     # D
-    term2 = (x2**2 - 7)**2 + 10*x2**2 - 45    # E
-
-    # term1 = himmelblau_hf([x1, x2])             # F
-    # term2 = x2**3 - (x1 + 1)**2                 # F
-
-    return term1 + term2
-
-
 def himmelblau_lf(xx):
     """
     HIMMELBLAU FUNCTION, LOWER FIDELITY CODE
@@ -80,14 +53,6 @@ def himmelblau_lf(xx):
 l_bound = [-4, -4]
 #: Upper bound for Himmelblau function
 u_bound = [ 4,  4]
-
-#: 2D Himmelblau function with fidelities 'high', 'medium' and 'low'
-himmelblau_3f = MultiFidelityFunction(
-    "himmelblau 3 fidelity",
-    u_bound, l_bound,
-    [himmelblau_hf, himmelblau_mf, himmelblau_lf],
-    fidelity_names=['high', 'medium', 'low']
-)
 
 #: 2D Himmelblau function with fidelities 'high' and 'low'
 himmelblau = MultiFidelityFunction(
