### 2020-06-21 da669a2
bugfix: removed imports of removed functions
