### 2020-06-21 fa06513
Added tests file for MultiFidelityFunction class
