### 2020-06-21 207a403
Added tests for (in)valid dimensionality in factory functions
