### 2020-06-21 ca704af
Added test of bi_fidelity_functions
