### 2020-06-21 3a32a47
Added type hint float for adjustable factory functions
