### 2020-06-22 c09e93e
Added zenodo DOI badge
