### 2020-06-22 5067700
fixed typo in README example
