### 2020-06-24 e736c9d
updated conda badge url to conda-forge
