### 2020-07-24 d116c26
rewrote regression_test.py using pytest-regressions
* input now generated at runtime with fixed seed
* removed create_regression_data.py file, functionality now automatic
* removed old input/output files, added new output files
* updated requirements
