### 2020-07-24 65e90a0
updated README to explain pytest-regressions workflow (fixes #3)
