### 2020-07-24 72bd50f
made 'run the tests' command explicit in README
