### 2020-07-24 8744e43
dev requirements now listed as extras_require in setup
