### 2020-07-24 ae3c0ae
updated travis script
