+++ .travis.yml
@@ -8,8 +8,7 @@ matrix:
       sudo: true
     - python: 3.8
 install:
-  - pip install .
-  - pip install -U -r requirements-dev.txt
+  - pip install .[dev]
   - pip install coverage coveralls
 script:
   - pytest tests/regression_test.py
