### 2020-07-24 3296feb
Quick fix typo setup.py
