### 2020-08-19 #2049 @arfon
> My sincerest apologies for the delays. A host of issues have conspired against me, including, most recently, being without power and internet for 10 days. I will devote time today and tomorrow to finalizing the review. Sorry once again, and thanks for your patience.

Wonderful, thanks for the update @zbeekman! Sounds like it has been a challenging time for you recently. We're very grateful for your time here 💖
