### 2020-08-20 #2049 @zbeekman
@sjvrijn I know this is not your fault or your responsibility, however the Simon Fraser University website links (with the matlab codes) that the documentation links to appear to all be broken. If you know of an updated URL that works, it would be nice to "fix" that, but I suspect it may just be unmaintained by the authors.
