### 2020-08-20 #2049 @zbeekman
@sjvrijn I will summarize and consolidate (in a comment) the tweaks etc. I'd like to see, but for now I'm going to keep enumerating them as they come up as their own individual comment.

I think the README and/or RTD should have a stronger **statement of need**. I think this could be achieved by succinctly, but explicitly describing the problem mf2 is solving, and the intended audience. You could grab text from the paper, or you could just write a sentence or two about this. Right now it says what it is, but is not as clear and compelling about what problem it solves.

For the **community guidelines** it's OK as it is now, but if you break that text out into a file named `CONTRIBUTING.md` GitHub will point new contributors to it when they want to open an issue or a pull request. I'm a big fan of issue templates and pull request templates too, but these are probably overkill for your package.
