### 2020-08-20 3f93129
Update RTD conda install to use conda-forge
