### 2020-08-20 c6434d6
Match RTD install instructions to readme for clone
