### 2020-08-20 #2049 @zbeekman
> @sjvrijn It looks like either the installation procedure for conda installs and from source via git clone are not identical between [the README](https://github.com/sjvrijn/mf2#installation) and the [Sphinx/RTD docs](https://mf2.readthedocs.io/en/latest/install.html). In particular, for conda, the channel specified is different between the documents, and the from source installation differs on the last line and there is an additional line telling you how to install requirements-dev.txt. Now, I'm not a pip expert, so they may be effectively equivalent, but it would be good to ensure the RTD/Sphinx documentation matches the README. If this is different for the release that will be cut when the JOSS paper is published, and the RTD documentation will be updated at that time, then, you need not address this comment.
