### 2020-08-20 #2049 @zbeekman
I have effectively concluded my review. Things look great! However, in my opinion, the following minor issues still need addressing before publication:

1. README / Documentation **statement of need**: It would be good to briefly improve this is in the readme and documentation to better adhere to JOSS review criteria. These explicitly state what the package does, but do not explicitly state what problem is being solved, how and in what context the authors anticipate it to be useful and who the intended audience is. A brief sentence or two could help to make the statement of need explicit rather than implicit.
2. Very minor inconsistencies in the installation instructions between the README and documentation. [I submitted a PR to correct this](https://github.com/sjvrijn/mf2/pull/6), and unless something I wrote is incorrect, I think that [the PR](https://github.com/sjvrijn/mf2/pull/6) should ideally be merged or the issue addressed by some other means before publication
3. *Optional*: Consider putting the contributing guidelines in a `CONTRIBUTING.md` file so that GitHub will point contributors in that direction before they open an issue or pull request.

Very nice work on the paper and the package, @sjvrijn! I tested both installation from PyPI and installation with development dependencies from a git clone on macOS and Linux. I ran all the regression and unit tests and reproduced the example figure. I also verified the scaling plots on a 2.9 GHz Intel Core i9 MacBook Pro with 6 cores.

Nice work once again and sorry my review was delayed. As soon as a sentence or two is added to the README and/or docs I'll tick off that last check box!
