### 2020-08-23 6ba68c5
small sentence improvement in JOSS paper
