### 2020-08-23 45e01dd
Improved README/docs intro: stronger statement of need
