### 2020-08-23 fa76a91
moved contribution guidelines to contributing.md
