+++ .github/ISSUE_TEMPLATE/function-suggestion.md
@@ -1,0 +1,20 @@
+---
+name: Function suggestion
+about: Suggest a new benchmark function to include
+title: "[Suggestion] <function(s)> from <author et al.>"
+labels: New function
+assignees: ''
+
+---
+
+**In which paper is this function introduced?**
+"Title", authors and DOI.
+
+**What function(s) should be added from this paper**
+Short name(s) of the function(s), as used in the paper.
+
+**What are characteristics of these functions?**
+Indicate briefly for each function their dimensionality, relevant landscape features, and any other peculiarities.
+
+**Additional information**
+Add any other relevant information here
