+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,30 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Code example to reproduce the behavior:
+```python
+import mf2
+
+...
+```
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Version information:**
+ - MF2 version:
+ - Numpy version:
+ - Python version:
+
+**Additional context**
+Add any other context about the problem here.
