+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,8 @@
+<!-- Feel free to remove check-list items aren't relevant to your change -->
+
+<!-- New function addition checklist -->
+ - [ ] New function has summary docstring with source paper DOI
+ - [ ] Summary added to `changelog.txt`
+ - [ ] Function added to `tests\regression_test.py` and `tests\property_test.py`
+ - [ ] Regression test files added
+ - [ ] Tests pass and coverage is not reduced
