### 2020-08-23 c8560e9
Added link to Gitter in issue template chooser
