+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,4 @@
+contact_links:
+  - name: Gitter chat room
+    url: https://gitter.im/pymf2/community
+    about: Please ask and answer questions here.
