### 2020-08-23 2f63c45
Minor enumeration spacing update
