### 2020-08-24 #2049 @zbeekman
> Would it be wise/useful to 'secure' the matlab source files somewhere else?

Since both projects are GPL, I personally would archive the MATLAB code in a folder with a readme describing when it was retrieved, from where, references etc. But I'm not saying you need to do that for a successful JOSS review.

> * Fair enough, I'll take a look at that in a bit and ping you when it's updated

It now looks perfect to me! I think this will help others who stumble upon the project to better understand how and why to use it!

> * Thanks for the PR, I've merged it 👍

Great, I always like to try to make things easy when I can contribute!

> * Good point, hadn't thought of that yet. Think I'll follow your suggestion there.

To be clear, this is __*NOT* __ blocking acceptance, so I'm happy whatever you decide. It looks like you did move it, and I'm going to open up a PR to make the reference in the README a link to the document.

All in all, great work @sjvrijn!

__*I fully recommend publication, no further modifications/edits required!*__
