### 2020-08-24 2d0acb5
Linkify reference to CONTRIBUTING.md from README.
