### 2020-08-24 #2049 @sjvrijn
> Since both projects are GPL, I personally would archive the MATLAB code in a folder with a readme describing when it was retrieved, from where, references etc. But I'm not saying you need to do that for a successful JOSS review.

That sounds like a decent solution to preserve the code. I'll add it to a subfolder in the [`docs/scripts`](https://github.com/sjvrijn/mf2/tree/master/docs/scripts) folder for completeness' sake.

> To be clear, this is ___NOT_ __ blocking acceptance, so I'm happy whatever you decide. It looks like you did move it, and I'm going to open up a PR to make the reference in the README a link to the document.

No worries, I realised it wasn't, but simply thought it was a good idea as I hadn't realised GitHub points to it when writing issues/PRs. Thanks again for fixing the link 👍

> All in all, great work @sjvrijn!
>
> **_I fully recommend publication, no further modifications/edits required!_**

Thanks for the review! 🎉
