### 2020-08-24 #2049 @arfon
@sjvrijn - I made a couple of minor changes in https://github.com/sjvrijn/mf2/pull/8 to your paper.

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
