+++ docs/scripts/matlab/curretal88explc.m
@@ -1,0 +1,50 @@
+function [y] = curretal88explc(xx)
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+% CURRIN ET AL. (1988) EXPONENTIAL FUNCTION, LOWER FIDELITY CODE
+% Calls: curretal88exp.m
+% This function, from Xiong et al. (2013), is used as the "low-accuracy
+% code" version of the function curretal88exp.m.
+%
+% Authors: Sonja Surjanovic, Simon Fraser University
+%          Derek Bingham, Simon Fraser University
+% Questions/Comments: Please email Derek Bingham at dbingham@stat.sfu.ca.
+%
+% Copyright 2013. Derek Bingham, Simon Fraser University.
+%
+% THERE IS NO WARRANTY, EXPRESS OR IMPLIED. WE DO NOT ASSUME ANY LIABILITY
+% FOR THE USE OF THIS SOFTWARE.  If software is modified to produce
+% derivative works, such modified software should be clearly marked.
+% Additionally, this program is free software; you can redistribute it 
+% and/or modify it under the terms of the GNU General Public License as 
+% published by the Free Software Foundation; version 2.0 of the License. 
+% Accordingly, this program is distributed in the hope that it will be 
+% useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+% of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+% General Public License for more details.
+%
+% For function details and reference information, see:
+% http://www.sfu.ca/~ssurjano/
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+% INPUT:
+%
+% xx = [x1, x2]
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+x1 = xx(1);
+x2 = xx(2);
+
+maxarg = max([0, x2-1/20]);
+
+yh1 = curretal88exp([x1+1/20, x2+1/20]);
+yh2 = curretal88exp([x1+1/20, maxarg]);
+yh3 = curretal88exp([x1-1/20, x2+1/20]);
+yh4 = curretal88exp([x1-1/20, maxarg]);
+
+y = (yh1 + yh2 + yh3 + yh4) / 4;
+
+end

[…]
