### 2020-08-24 555ead5
bumped version to 2020.8.0
